Moved unit tests to a separate dir
