+++ ydb/core/kqp/ut/new_rbo/ya.make
@@ -1,0 +1,35 @@
+UNITTEST_FOR(ydb/core/kqp)
+
+FORK_SUBTESTS()
+
+SIZE(MEDIUM)
+
+SRCS(
+    kqp_newrbo_ut.cpp
+)
+
+PEERDIR(
+    ydb/core/kqp/ut/common
+    yql/essentials/public/udf/service/exception_policy
+    yql/essentials/sql/pg
+    yql/essentials/parser/pg_wrapper
+    ydb/public/lib/ut_helpers
+)
+
+ADDINCL(
+    yql/essentials/parser/pg_wrapper/postgresql/src/include
+)
+
+IF (OS_WINDOWS)
+CFLAGS(
+   "-D__thread=__declspec(thread)"
+   -Dfstat=microsoft_native_fstat
+   -Dstat=microsoft_native_stat
+)
+ENDIF()
+
+NO_COMPILER_WARNINGS()
+
+YQL_LAST_ABI_VERSION()
+
+END()
