+++ ydb/core/kqp/ut/pg/kqp_pg_ut.cpp
@@ -11,8 +11,6 @@
 #include <ydb/public/lib/ut_helpers/ut_helpers_query.h>
 #include <util/system/env.h>
 
-#include <ctime>
-
 
 extern "C" {
 #include "catalog/pg_type_d.h"
@@ -40,44 +38,6 @@ namespace NKqp {
 
 using namespace NYdb;
 using namespace NYdb::NTable;
-
-double TimeQuery(NKikimr::NKqp::TKikimrRunner& kikimr, TString query, int nIterations) {
-    auto db = kikimr.GetTableClient();
-    auto session = db.CreateSession().GetValueSync().GetSession();
-
-    clock_t the_time;
-    double elapsed_time;
-    the_time = clock();
-
-    for (int i=0; i<nIterations; i++) {
-        //session.ExecuteDataQuery(query, TTxControl::BeginTx().CommitTx()).GetValueSync();
-        session.ExplainDataQuery(query).GetValueSync();
-    }
-
-    elapsed_time = double(clock() - the_time) / CLOCKS_PER_SEC;
-    return elapsed_time / nIterations;        
-}
-
-double TimeQuery(TString schema, TString query, int nIterations) {
-    NKikimrConfig::TAppConfig appConfig;
-    appConfig.MutableTableServiceConfig()->SetEnableNewRBO(true);
-    TKikimrRunner kikimr(NKqp::TKikimrSettings().SetWithSampleTables(false).SetAppConfig(appConfig));
-    auto db = kikimr.GetTableClient();
-    auto session = db.CreateSession().GetValueSync().GetSession();
-    session.ExecuteSchemeQuery(schema).GetValueSync();
-
-    clock_t the_time;
-    double elapsed_time;
-    the_time = clock();
-
-    for (int i=0; i<nIterations; i++) {
-        //session.ExecuteDataQuery(query, TTxControl::BeginTx().CommitTx()).GetValueSync();
-        session.ExplainDataQuery(query).GetValueSync();
-    }
-
-    elapsed_time = double(clock() - the_time) / CLOCKS_PER_SEC;
-    return elapsed_time / nIterations;        
-}
 
 NYdb::NTable::TDataQueryResult ExecutePgSelect(
     NKikimr::NKqp::TKikimrRunner& kikimr, const TString& tableName)
@@ -1177,411 +1137,6 @@ Y_UNIT_TEST_SUITE(KqpPg) {
         CompareYson(R"([
             ["1";"2";"12"]
         ])", FormatResultSetYson(result.GetResultSet(0)));
-    }
-
-    Y_UNIT_TEST(NewRBO1) {
-        NKikimrConfig::TAppConfig appConfig;
-        appConfig.MutableTableServiceConfig()->SetEnableNewRBO(true);
-        TKikimrRunner kikimr(NKqp::TKikimrSettings().SetWithSampleTables(false).SetAppConfig(appConfig));
-        auto db = kikimr.GetTableClient();
-        auto session = db.CreateSession().GetValueSync().GetSession();
-
-        auto result = session.ExecuteDataQuery(R"(
-            --!syntax_pg
-            SELECT 1 as "a", 2 as "b";
-        )", TTxControl::BeginTx().CommitTx()).GetValueSync();
-
-        UNIT_ASSERT_C(result.IsSuccess(), result.GetIssues().ToString());
-    }
-
-    Y_UNIT_TEST(NewRBO2) {
-        NKikimrConfig::TAppConfig appConfig;
-        appConfig.MutableTableServiceConfig()->SetEnableNewRBO(true);
-        TKikimrRunner kikimr(NKqp::TKikimrSettings().SetWithSampleTables(false).SetAppConfig(appConfig));
-        auto db = kikimr.GetTableClient();
-        auto session = db.CreateSession().GetValueSync().GetSession();
-
-        session.ExecuteSchemeQuery(R"(
-            CREATE TABLE `/Root/foo` (
-                id	Int64	NOT NULL,	
-	            name	String,
-                primary key(id)	
-            );
-        )").GetValueSync();
-
-        db = kikimr.GetTableClient();
-        auto session2 = db.CreateSession().GetValueSync().GetSession();
-
-        auto result = session2.ExecuteDataQuery(R"(
-            --!syntax_pg
-            SET TablePathPrefix = "/Root/";
-            SELECT id as "id2" FROM foo WHERE name = 'some_name';
-        )", TTxControl::BeginTx().CommitTx()).GetValueSync();
-
-        UNIT_ASSERT_C(result.IsSuccess(), result.GetIssues().ToString());
-    }
-
-    Y_UNIT_TEST(NewRBO3) {
-        NKikimrConfig::TAppConfig appConfig;
-        appConfig.MutableTableServiceConfig()->SetEnableNewRBO(true);
-        TKikimrRunner kikimr(NKqp::TKikimrSettings().SetWithSampleTables(false).SetAppConfig(appConfig));
-        auto db = kikimr.GetTableClient();
-        auto session = db.CreateSession().GetValueSync().GetSession();
-
-        session.ExecuteSchemeQuery(R"(
-            CREATE TABLE `/Root/foo` (
-                id	Int64	NOT NULL,	
-	            name	String,
-                primary key(id)	
-            );
-
-            CREATE TABLE `/Root/bar` (
-                id	Int64	NOT NULL,	
-	            lastname	String,
-                primary key(id)	
-            );
-        )").GetValueSync();
-
-        db = kikimr.GetTableClient();
-        auto session2 = db.CreateSession().GetValueSync().GetSession();
-
-        auto result = session2.ExecuteDataQuery(R"(
-            --!syntax_pg
-            SET TablePathPrefix = "/Root/";
-            SELECT f.id as "id2" FROM foo AS f, bar WHERE name = 'some_name';
-        )", TTxControl::BeginTx().CommitTx()).GetValueSync();
-
-        UNIT_ASSERT_C(result.IsSuccess(), result.GetIssues().ToString());
-    }
-
-    Y_UNIT_TEST(NewRBO4) {
-        NKikimrConfig::TAppConfig appConfig;
-        appConfig.MutableTableServiceConfig()->SetEnableNewRBO(true);
-        TKikimrRunner kikimr(NKqp::TKikimrSettings().SetWithSampleTables(false).SetAppConfig(appConfig));
-        auto db = kikimr.GetTableClient();
-        auto session = db.CreateSession().GetValueSync().GetSession();
-
-        session.ExecuteSchemeQuery(R"(
-            CREATE TABLE `/Root/foo` (
-                id	Int64	NOT NULL,	
-	            name	String,
-                primary key(id)	
-            );
-
-            CREATE TABLE `/Root/bar` (
-                id	Int64	NOT NULL,	
-	            lastname	String,
-                primary key(id)	
-            );
-        )").GetValueSync();
-
-        db = kikimr.GetTableClient();
-        auto session2 = db.CreateSession().GetValueSync().GetSession();
-
-        auto result = session2.ExecuteDataQuery(R"(
-            --!syntax_pg
-            SET TablePathPrefix = "/Root/";
-            SELECT f.id as "id2" FROM foo AS f, bar WHERE f.id = bar.id and name = 'some_name';
-        )", TTxControl::BeginTx().CommitTx()).GetValueSync();
-
-        UNIT_ASSERT_C(result.IsSuccess(), result.GetIssues().ToString());
-    }
-
-    Y_UNIT_TEST(Bench_NewRBO1) {
-        NKikimrConfig::TAppConfig appConfig;
-        appConfig.MutableTableServiceConfig()->SetEnableNewRBO(true);
-        TKikimrRunner kikimr(NKqp::TKikimrSettings().SetWithSampleTables(false).SetAppConfig(appConfig));
-
-        auto time = TimeQuery(kikimr, R"(
-                --!syntax_pg
-                SELECT 1 as "a", 2 as "b";
-            )", 1000);
-
-        Cout << "Time per query: " << time;     
-
-        //UNIT_ASSERT_C(result.IsSuccess(), result.GetIssues().ToString());
-    }
-
-    Y_UNIT_TEST(Bench_NewRBO2) {
-        NKikimrConfig::TAppConfig appConfig;
-        appConfig.MutableTableServiceConfig()->SetEnableNewRBO(true);
-        TKikimrRunner kikimr(NKqp::TKikimrSettings().SetWithSampleTables(false).SetAppConfig(appConfig));
-        auto db = kikimr.GetTableClient();
-        auto session = db.CreateSession().GetValueSync().GetSession();
-
-        session.ExecuteSchemeQuery(R"(
-            CREATE TABLE `/Root/foo` (
-                id	Int64	NOT NULL,	
-	            name	String,
-                primary key(id)	
-            );
-        )").GetValueSync();
-
-        auto time = TimeQuery(kikimr, R"(
-            --!syntax_pg
-            SET TablePathPrefix = "/Root/";
-            SELECT id as "id2" FROM foo WHERE name = 'some_name';
-        )",1000);
-
-        Cout << "Time per query: " << time;     
-    }
-
-    Y_UNIT_TEST(Bench_NewRBO3) {
-        NKikimrConfig::TAppConfig appConfig;
-        appConfig.MutableTableServiceConfig()->SetEnableNewRBO(true);
-        TKikimrRunner kikimr(NKqp::TKikimrSettings().SetWithSampleTables(false).SetAppConfig(appConfig));
-        auto db = kikimr.GetTableClient();
-        auto session = db.CreateSession().GetValueSync().GetSession();
-
-        session.ExecuteSchemeQuery(R"(
-            CREATE TABLE `/Root/foo` (
-                id	Int64	NOT NULL,	
-	            name	String,
-                primary key(id)	
-            );
-
-            CREATE TABLE `/Root/bar` (
-                id	Int64	NOT NULL,	
-	            lastname	String,
-                primary key(id)	
-            );
-        )").GetValueSync();
-
-        auto time = TimeQuery(kikimr, R"(
-            --!syntax_pg
-            SET TablePathPrefix = "/Root/";
-            SELECT f.id as "id2" FROM foo AS f, bar WHERE name = 'some_name';
-        )", 1000);
-
-        Cout << "Time per query: " << time;     
-    }
-
-    Y_UNIT_TEST(Bench_NewRBO4) {
-        NKikimrConfig::TAppConfig appConfig;
-        appConfig.MutableTableServiceConfig()->SetEnableNewRBO(true);
-        TKikimrRunner kikimr(NKqp::TKikimrSettings().SetWithSampleTables(false).SetAppConfig(appConfig));
-        auto db = kikimr.GetTableClient();
-        auto session = db.CreateSession().GetValueSync().GetSession();
-
-        session.ExecuteSchemeQuery(R"(
-            CREATE TABLE `/Root/foo` (
-                id	Int64	NOT NULL,	
-	            name	String,
-                primary key(id)	
-            );
-
-            CREATE TABLE `/Root/bar` (
-                id	Int64	NOT NULL,	
-	            lastname	String,
-                primary key(id)	
-            );
-        )").GetValueSync();
-
-        auto time = TimeQuery(kikimr, R"(
-            --!syntax_pg
-            SET TablePathPrefix = "/Root/";
-            SELECT f.id as "id2" FROM foo AS f, bar WHERE f.id = bar.id and name = 'some_name';
-        )", 1000);
-
-        Cout << "Time per query: " << time;     
-    }
-
-    Y_UNIT_TEST(Bench_NewRBO5) {
-        NKikimrConfig::TAppConfig appConfig;
-        appConfig.MutableTableServiceConfig()->SetEnableNewRBO(true);
-        TKikimrRunner kikimr(NKqp::TKikimrSettings().SetWithSampleTables(false).SetAppConfig(appConfig));
-        auto db = kikimr.GetTableClient();
-        auto session = db.CreateSession().GetValueSync().GetSession();
-
-        auto schema = R"(
-CREATE TABLE `/Root/foo_0` (
-    id Int64 NOT NULL,
-    join_id Int64,
-    primary key(id)
-    );
-    
-
-    CREATE TABLE `/Root/foo_1` (
-    id Int64 NOT NULL,
-    join_id Int64,
-    primary key(id)
-    );
-    
-
-    CREATE TABLE `/Root/foo_2` (
-    id Int64 NOT NULL,
-    join_id Int64,
-    primary key(id)
-    );
-    
-
-    CREATE TABLE `/Root/foo_3` (
-    id Int64 NOT NULL,
-    join_id Int64,
-    primary key(id)
-    );
-    
-
-    CREATE TABLE `/Root/foo_4` (
-    id Int64 NOT NULL,
-    join_id Int64,
-    primary key(id)
-    );
-    
-
-    CREATE TABLE `/Root/foo_5` (
-    id Int64 NOT NULL,
-    join_id Int64,
-    primary key(id)
-    );
-    
-
-    CREATE TABLE `/Root/foo_6` (
-    id Int64 NOT NULL,
-    join_id Int64,
-    primary key(id)
-    );
-    
-
-    CREATE TABLE `/Root/foo_7` (
-    id Int64 NOT NULL,
-    join_id Int64,
-    primary key(id)
-    );
-    
-
-    CREATE TABLE `/Root/foo_8` (
-    id Int64 NOT NULL,
-    join_id Int64,
-    primary key(id)
-    );
-    
-
-    CREATE TABLE `/Root/foo_9` (
-    id Int64 NOT NULL,
-    join_id Int64,
-    primary key(id)
-    );
-    )";
-
-        auto query = R"(
-            --!syntax_pg
-     SET TablePathPrefix = "/Root/";
-
-     SELECT foo_0.id as "id2"
-     FROM foo_0, foo_1, foo_2, foo_3, foo_4, foo_5, foo_6, foo_7, foo_8, foo_9
-     WHERE foo_0.join_id = foo_1.id AND foo_0.join_id = foo_2.id AND foo_0.join_id = foo_3.id AND foo_0.join_id = foo_4.id AND foo_0.join_id = foo_5.id AND foo_0.join_id = foo_6.id AND foo_0.join_id = foo_7.id AND foo_0.join_id = foo_8.id AND foo_0.join_id = foo_9.id;
-     
-    )";
-
-        auto time = TimeQuery(schema, query, 100);
-
-        Cout << "Time per query: " << time;     
-    }
-
-    Y_UNIT_TEST(YQL_NewRBO1) {
-        NKikimrConfig::TAppConfig appConfig;
-        appConfig.MutableTableServiceConfig()->SetEnableNewRBO(true);
-        TKikimrRunner kikimr(NKqp::TKikimrSettings().SetWithSampleTables(false).SetAppConfig(appConfig));
-        auto db = kikimr.GetTableClient();
-        auto session = db.CreateSession().GetValueSync().GetSession();
-
-        auto result = session.ExecuteDataQuery(R"(
-            SELECT 1 as a, 2 as b;
-        )", TTxControl::BeginTx().CommitTx()).GetValueSync();
-
-        UNIT_ASSERT_C(result.IsSuccess(), result.GetIssues().ToString());
-    }
-
-    Y_UNIT_TEST(YQL_NewRBO2) {
-        NKikimrConfig::TAppConfig appConfig;
-        appConfig.MutableTableServiceConfig()->SetEnableNewRBO(true);
-        TKikimrRunner kikimr(NKqp::TKikimrSettings().SetWithSampleTables(false).SetAppConfig(appConfig));
-        auto db = kikimr.GetTableClient();
-        auto session = db.CreateSession().GetValueSync().GetSession();
-
-        session.ExecuteSchemeQuery(R"(
-            CREATE TABLE `/Root/foo` (
-                id	Int64	NOT NULL,	
-	            name	String,
-                primary key(id)	
-            );
-        )").GetValueSync();
-
-        db = kikimr.GetTableClient();
-        auto session2 = db.CreateSession().GetValueSync().GetSession();
-
-        auto result = session2.ExecuteDataQuery(R"(
-            PRAGMA TablePathPrefix = "/Root/";
-            SELECT id as id2 FROM foo WHERE name = 'some_name';
-        )", TTxControl::BeginTx().CommitTx()).GetValueSync();
-
-        UNIT_ASSERT_C(result.IsSuccess(), result.GetIssues().ToString());
-    }
-
-    Y_UNIT_TEST(YQL_NewRBO3) {
-        NKikimrConfig::TAppConfig appConfig;
-        appConfig.MutableTableServiceConfig()->SetEnableNewRBO(true);
-        TKikimrRunner kikimr(NKqp::TKikimrSettings().SetWithSampleTables(false).SetAppConfig(appConfig));
-        auto db = kikimr.GetTableClient();
-        auto session = db.CreateSession().GetValueSync().GetSession();
-
-        session.ExecuteSchemeQuery(R"(
-            CREATE TABLE `/Root/foo` (
-                id	Int64	NOT NULL,	
-	            name	String,
-                primary key(id)	
-            );
-
-            CREATE TABLE `/Root/bar` (
-                id	Int64	NOT NULL,	
-	            lastname	String,
-                primary key(id)	
-            );
-        )").GetValueSync();
-
-        db = kikimr.GetTableClient();
-        auto session2 = db.CreateSession().GetValueSync().GetSession();
-
-        auto result = session2.ExecuteDataQuery(R"(
-            PRAGMA TablePathPrefix = "/Root/";
-            SELECT f.id as id2 FROM foo AS f CROSS JOIN bar WHERE name = 'some_name';
-        )", TTxControl::BeginTx().CommitTx()).GetValueSync();
-
-        UNIT_ASSERT_C(result.IsSuccess(), result.GetIssues().ToString());
-    }
-
-    Y_UNIT_TEST(YQL_NewRBO4) {
-        NKikimrConfig::TAppConfig appConfig;
-        appConfig.MutableTableServiceConfig()->SetEnableNewRBO(true);
-        TKikimrRunner kikimr(NKqp::TKikimrSettings().SetWithSampleTables(false).SetAppConfig(appConfig));
-        auto db = kikimr.GetTableClient();
-        auto session = db.CreateSession().GetValueSync().GetSession();
-
-        session.ExecuteSchemeQuery(R"(
-            CREATE TABLE `/Root/foo` (
-                id	Int64	NOT NULL,	
-	            name	String,
-                primary key(id)	
-            );
-
-            CREATE TABLE `/Root/bar` (
-                id	Int64	NOT NULL,	
-	            lastname	String,
-                primary key(id)	
-            );
-        )").GetValueSync();
-
-        db = kikimr.GetTableClient();
-        auto session2 = db.CreateSession().GetValueSync().GetSession();
-
-        auto result = session2.ExecuteDataQuery(R"(
-            PRAGMA TablePathPrefix = "/Root/";
-            PRAGMA ydb.OptOverrideStatistics = '{"/Root/foo": {"n_rows": 1e9}, "/Root/bar": {"n_rows" : 1e9}}';
-            SELECT f.id as id2 FROM foo AS f CROSS JOIN bar WHERE f.id = bar.id and name = 'some_name';
-        )", TTxControl::BeginTx().CommitTx()).GetValueSync();
-
-        UNIT_ASSERT_C(result.IsSuccess(), result.GetIssues().ToString());
     }
 
     Y_UNIT_TEST(ReadPgArray) {
